requirements.txt
matplotlib
tensorflow-gpu==1.4.1
scipy==1.0.0
opencv-python==3.3.0.10
numpy==1.11.0
Pillow==5.1.0
